bug (Leak)

+++ src/error.c
@@ -6,7 +6,7 @@
 /*   By: mbrousse <mbrousse@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/09 16:05:08 by mbrousse          #+#    #+#             */
-/*   Updated: 2024/03/06 18:41:11 by mbrousse         ###   ########.fr       */
+/*   Updated: 2024/03/06 19:44:37 by mbrousse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -68,7 +68,7 @@ int	ft_error_duplicated(t_stack_node **list, int nbr)
 
 void	ft_freetab(char **tab)
 {
-	int	i;
+	size_t	i;
 
 	if (!tab)
 		return ;
